+++ webapp_frontend/src/components/pages/errors/Error404.stories.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import {storiesOf} from "@storybook/react";
+import Error404 from "./Error404";
+
+storiesOf("ColorButton", module)
+    .add("default", () => <Error404/>
+    )
